Move manifest store to pkg/kube/util
[#163483576](https://www.pivotaltracker.com/story/show/163483576)
