+++ pkg/kube/util/store/manifest/manifest.go
