+++ pkg/kube/util/store/manifest/manifest_suite_test.go
